backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
ollama
httpx==0.27.0
pytest>=8.0.0
